+++ tests/test_molecule.py
@@ -208,3 +208,16 @@ def test_split_molecule(water_u: "Universe") -> None:
     assert len(Chem.GetMolFrags(pmol)) == 1
     assert len(Chem.GetMolFrags(wmol)) == 11
     assert combined["TIP310.5"] is wmol["TIP310.5"]
+
+
+def test_split_molecule_reassigns_mapindex(u: "Universe") -> None:
+    """split_molecule passes the original residues to the split mols, but needs to
+    reset the mapindex first, otherwise the mapindex appearing in residues doesn't match
+    with the atom in the parent molecule or can result in out of bound indices,
+    typically when the LHS is further in the input mol than the RHS"""
+    peptide = u.select_atoms("protein and resid 40-50")
+    system_mol = Molecule.from_mda(peptide)
+    lhs, _rhs = split_molecule(system_mol, lambda x: x.number == 49)
+    assert lhs.n_residues == 1
+    # if not reset, this would be 148
+    assert lhs[0].GetAtomWithIdx(0).GetUnsignedProp("mapindex") == 0